+++ synthetic_data/distinct_generators/str_generator.py
